fix: ディレクトリ構成の見直し
